requirements.txt
pyautogui
schedule
pystray
Pillow
customtkinter
